Allow Suites to be added to an Engine over time.
Instead of providing all test suites to the Engine's constructor, the
engine now has a sink that can be used to add suites over time and
closed to indicate that no more suites are coming. This makes the use
pattern more flexible, and is a step towards supporting a live test
runner and to more memory efficiency.

Reporters no longer wrap the Engine. Instead, they expose a static
`watch` function which subscribes to the events the Engine emits.

R=kevmoo@google.com

Review URL: https://codereview.chromium.org//1187103004.