Factor out a Runner class from executable.dart.
This is useful as an object because it provides all the private methods with
full access to the configuration, the engine, and the loader without having to
pass them around or create top-level variables. It also creates a cleaner
distinction between receiving a signal and closing various components.

R=kevmoo@google.com

Review URL: https://codereview.chromium.org//1239613002 .