Use FIFO timer ordering in FakeAsync (dart-lang/fake_async#265)
As noted in https://github.com/google/quiver-dart/issues/258, the
current ordering of _getNextTimer() in _FakeAsync is inconsistent with
the native Dart event loop.

In the process of fixing that, I had to fix a few other idiosyncrasies
that were revealed from the existing unit tests which relied on the
incorrect behavior. For instance, if you had a periodic timer occur at
the same time as a delayed timer, whether or not both of them were
"flushed" in flushTimers depended on the order with which they were
queued. Now it does not.