Support re-running tests while debugging.
When the browser is paused at a breakpoint, the runner will now open a
command-line console with a "restart" command that will restart the
current test.

This currently has no tests, due to dart-lang/sdk#25369 and the lack of
a programmatic API for interacting with Observatory. I've tested it
reasonably thoroughly by hand, but it's more likely than the average
feature to have lurking bugs.

Closes #335

R=kevmoo@google.com

Review URL: https://codereview.chromium.org//1561073003 .