Add configuration support for choosing test paths.
This adds two fields: "paths" allows users to specify the paths that are
used if nothing is passed on the command line, and "filename" allows
users to declare which individual files should be loaded.

I'm closing out the package-level config file issue because this is the
last non-plugin feature mentioned in it. There's still a lot more work
to do, but that's tracked in separate issues.

Closes #46

R=kevmoo@google.com

Review URL: https://codereview.chromium.org//1668573003 .