Add a platform plugin infrastructure.
This is basically the minimum Flutter needs to stop doing manual
serialization stuff. This also migrates the VM loader to use the plugin
infrastructure.

In a follow-up commit, I plan to migrate the browser loader as well.
That will likely require some additional modifications to the plugin API
to support the more complex use-case.

R=kevmoo@google.com

Review URL: https://codereview.chromium.org//1685363002 .