Default to using the system packageRoot.
Currently, the default is to use 'pacakges/' in the cwd as the package
root. That means if you set a different packageRoot as a flag to the
dart vm, you also need to set it as a flag to package:test. To simplify
things, package:test can just use whatever the Platform packageRoot is
by default, so you only need to set it once.