Add metadata to messages emitted by live tests.
This moves the responsibility for printing skip messages to the engine,
rather than the reporter. Reporters now just need to choose how to
display the messages produced by the engine.

This lays the groundwork for #394, since skip messages can now be
emitted during the run of a test.

R=kevmoo@google.com

Review URL: https://codereview.chromium.org//2091173002 .