Run tear-downs in the same error zone as tests. (#643)
Run tear-downs in the same error zone as tests.

We used to create a new error zone for each call to
waitForOutstandingCallbacks(), which meant that asynchronous errors
couldn't be safely passed around within a single test. Now we have a
single error handler that manually checks the zone in which the error
was thrown.