Replace implicit dynamic with Null in arity checks (#675)
Fixes #674

dynamic is treated as the bottom type for function arguments, but this
will soon change. Null will continue to be the bottom type.

The question we are asking is "Is this a function that takes one
argument", when dynamic is no longer the bottom type the question will
become "Is this a function that takes one dynamic argument", and a
function which take one argument of a specific type does *not* take a
dynamic argument.