More Dart 2 runtime semantic fixes
```
type 'WebSocketChannel' is not a subtype of type 'StreamChannel<String>' of 'channel'
```

Cast the `WebSocketChannel` to indicate that all values in each
direction will be `String`.

```
type 'CloseGuaranteeChannel<dynamic>' is not a subtype of type 'VirtualChannel'
```

Avoid using the `suiteChannel` variable for two types. It needs to start
as a `VirtualChannel` which has the `id` field, and then later be a
`StreamChannel`.

```
type 'RunnerSuite' is not a subtype of type 'FutureOr<Future<RunnerSuite>>'
```

Avoid the attempted Future flattenting which was never correct but was
supported in the Dart 1 VM.