Drop AsyncFunction typedef
The typedef was misused to mean either `Function` or `Function()` in
various places and worked because the difference was not checked. The
doc comment was also incorrect because there were places it was used
where a return value was explicitly necessary. Replace with an inline
function type definition using the appropriate choice for the location.

Where possible also replace with Function types that flow through the
required return types.