Fix implicit casts in lib directory
There are still errors in tests which may be harder to fix so we can't
check in the change to analysis options.

This should make it easier to diagnose Dart 2 semantic issues as more of
the places where there can be failures are explicit.

Most of the changes are just adding explicit `as` statements and keeping
the semantics that already existed. In a few places an uninitialized
`var` is given a type, and in others a reused variable name with
different types is split into multiple variables.

The changes in the frontend code are the most risky, since some users
might still be using that code in a non Dart 2 runtime and relying on
looseness which is lost with an explicit cast.