Add types to a bunch of uninitalized variables
Searched for lines matching `var \w+;`. This is a much weaker version of
no-implicit-dynamic. This catches a bunch of places where we otherwise
had dynamic calls and tightens up types a bit. Where we are relying on
being able to parse to multiple types and then throw a specific
excpetion or return null on bad types I added an explicit `dynamic`.