Re-enable some skipped tests for a stale issue (#870)
We no longer use `mapMap` so the hints aren't printed. Fix the tests to
allow for the "Compiled" message from dart2js. Update the reason for
skipping one test which has an incorrect line number - we will need to
investigate why this line is wrong.

Allow Matcher usage in the `_print` call since the "Compiled" message is
unpredictable.