Allow for multiple calls to executable.main()
In Flutter, we would like to have our test runner be able to watch
a directory and re-run tests on file changes.  This requires making
multiple calls to the test_core executable, which with current
behavior causes it to throw on multiple calls to listen() on the
process signals stream.

Since the stream is properly cleaned up between runs, and we ensure
that only one subscription is active, move the _signals stream from
a final variable to a getter lambda expression.

Verified that the behavior is stable across multiple test runs within
a single process