Stop tagging `MessageChannel` messages with href
Messages sent from iframes to the host via `window.postMessage()` are
tagged with the iframe's href so that the host can associate the message
with the sending iframe. This is necessary because messages sent from
all iframes are received by the same event handler.

However, this isn't necessary for messages sent over a `MessageChannel`,
as the host uses a separate `MessageChannel` for each iframe.