Remove unnecessary DelegatingStreamSink wrapping
This was added in https://github.com/dart-lang/test/pull/729 when the
`Socket` class was missing a method in an interface it declared it had -
this caused a problem when some code tried to tear if off.

According to https://github.com/dart-lang/sdk/issues/31504 that method
now exists and can be torn off - calling that method would still throw,
but that would also be true with a `DelegatignStreamSink` wrapper.