Remove uses of "invoke" utility (#1119)
This has the same intended purpose as `unawaited` so it can be easily
replaced. Technically even the `unawaited` is not necessary here since
none of the surrounding functions are `async`, but it is fine to use it
for documentation purposed.

Leave the utility in place for now since it will take some version bump
gymnastics to roll out.