Always pass a function to BooleanSelector.evaluate
Prepare for a potential breaking change in `package:boolean_selector`.
Currently the `evaluate` method takes an argument which is either an
`Iterable<String>` or a `bool Function(String)` which matches the
`.toSet()contains` tearoff on that Iterable. We might tighten that to
always take a `bool Function(String)` since the tearoff is easy to
handle at the calling side.