Drop optional new from code in strings
Remove the `new` from Dart code in strings, docs, comments, and the
README. Leave the keyword if it's used in a doc comment references to
link specifically to the unnamed constructor instead of a class.

Simplify some implementation comments to avoid repeating information.

In some user facing Dart code, switch to single quotes for consistency.