Avoid unhandled errors in spawnHybridUri
Closes #1007
Closes #1008

We already forward errors across the `StreamChannel` so it isn't
particularly useful to also have them printed to the output on the VM
hosting the runner.

We know that `listen` is only ever called in an Isolate that was spawned
from the runner and the `zone` is the root zone. The uncaught error
handler on the root zone will print the error and then kill the isolate,
we jump straight to killing the isolate instead. This Isolate is only
ever spawned without an `onError` handler that would override the
default.