Remove an unnecessary await on event loop (#1164)
Closes #865

Move a `pumpEventQueue` to fix a failing test.

The test was passing because of the extra event loop cycle in
`close()`. The intent of the test is to check behavior when a test
hasn't completed yet. By closing immediately the test was actually
looking at behavior of when the tests hadn't even been declared yet, and
the extra `await` was hiding that.

Reorder the closing of the `_suiteController` and
`_onSuiteAddedController`. Values flow first through the
`_suitController` and then to the `_onSuitAddedController`. I saw at
least one case where a value flowed through late and cause an error
attempting to add a value on a closed controller.