Exit as a failure if no tests ran
Closes #764

If there were no passing tests at all exit with a failing exit code.
This could happen if all tests are skipped, or file is run through
the test runner that does not define any tests. This will reduce the
chances of unintentionally configuring CI such that it doesn't actually
check anything.