Re-enable skipped tests around stack traces
These were disabled because non-hermetic URIs were present. They have
been replaced with `org-dartlang-sdk` URIs. This does surface the fact
that there are UX differences from before, but we can at least reliably
test behavior now.

The tests in `pub_serve_test.dart` are still skipped for other reasons
so we can't guarantee they are fixed correctly here. Once we unskip
those tests we wouldn't want to hide issues with the stack trace
mapping, so we'll remove that reason for the skip.