Simplify currentOS initialization (#1221)
The previous code was working around the possibility of getting a `null`
OS, that that is never returned. Instead `OperatingSystem.none` is used.
Since we can't hit the null branch and exception, the entire initializer
can simplify to the static method call. This will be more clearly
correct after the null safety migration.