Improve docs for async matchers (#1225)
Closes #1224

Remove the last sentence which references `AsyncMatcher` by name since
that is an implementation detail, and the reference to `expect` was
confusing and contradicted the paragraph above.

Rephrase the paragraph about asynchronous matching to reference both
`expect` and `expectLater`. Expand on "the actual expectation" by
referencing the `matcher` argument explicitly.