Merge _LiveTest into LiveTestController
The `_LiveTest` was already a very thin wrapper around
`LiveTestController` which exposed some of it's private members.
Instead, make `LiveTestController` implement `LiveTest` directly and
skip the extra level of classes that have cyclical references.

This is a small refactoring to make it easier to understand structure
and dependencies for future changes.