Fail with specific error for null runnerSuite (#1245)
See https://github.com/flutter/flutter/issues/56457

If we get a null `runnerSuite` the attempt to read the `environment`
file will fail, resulting in a confusing error and stack trace. We can't
reproduce this locally so it isn't clear where the null may be coming
from.

Add additional information to the error thrown to try to narrow down the
places we need to search for a root cause.