MockBuilder: Widen the type of non-nullable parameters to be nullable.
The mechanism for doing this is to add some `forceNullable` parameters here and there, to ensure that overriding method parameters and overriding setter parameters are nullable. Other parameters, like those of function-typed parameters or generic function types, are not widened.

Include plenty of tests, and update all test expectations to now include widened parameters.

Update a few tests as well to be simpler, including only relevant content, and testing only relevant content.

PiperOrigin-RevId: 315296150