Remove dependency on `http_multi_socket`.
Closes #1264

Listening on both IPv6 and IPv4 allows for the case where `node` only
supports IPv4 but the OS supports IPv6, however this is not a situation
that should come up in the real world.

Bind to IPv6 and fall back to IPv4 if it doesn't work. The node side
calls `net.connect(port)` which should support either assuming OS
support.