Run tests in declaring zone
Three problems (at least) surfaced by using the Zone from when main was
run rather than the zone when the runner is asking the invoker to run:

1 The heartbeat stuff could get stuck by a FakeAsync zone when the users
  calls `test()`. It wouldn't let the timer run.
2 The communication channel for "remote" tests can't find the channel
  for the specific test that it should be speaking over.
3 We expect an Invoker.guard() to funnel the errors to the correct test
  case. If they surface in the same zone as we ran `main()` it looks
  like a test error.

For 1 I _think_ we can run the timer in the root zone since that should
always have sane async behavior. The invoker zone is necessarily a child
of it anyway.

For 2 With some interface changes we can pass the important zone
variable through rather than setting it on a zone in the remote
listener. There might also be a better overall structure that would make
this less awkward.

For 3 it seems sufficient to always "guard" the tests so that uncaught
errors show up for the right invoker, but there is some conversation
about letting errors (probably `Future.error`s...) cross from `setUpAll`
to `tearDownAll` without getting stuck in an error Zone. It _looks_ like
`Invoker.guard` should already handle some of that.