Ignore a null runner suite in json reporter
The `StateError` had been put in with #1245 in order to facilitate
debugging this problem when it surfaces sometimes for flutter.

In #1258 Jake pointed out that there is a comment which mentions that
the `runnerSuite` may be null if it is unavailable due to a load error.

I'm still not sure exactly the conditions that lead to this behavior,
but it should ideally come with some other more clear error, so ignoring
it here should hopefully make things less confusing.