Restore _guardIfGuarded
Without this an erroring future that originates in `setUpAll` can't
complete (to an error) in a teardown added by the setup with
`addTearDown`.

Instead use `Invoker.guard` and store the variable, Run both the `main`
and the tests themselves in that same zone. There might be a more
targeted way we can handle this though, and hopefully also alleviate the
`guardIfGuarded` stuff.