Refactor tests for consistent error zones
Previously the tests validated that an error Future which originates in
one zone is able to complete as an error in the other zone. Since errors
are not allowed to cross zone boundaries the future would never complete
and result in a test timeout.

Zones can indicate directly when they are in the same error zone. Using
this API means the test can fail fast instead of a waiting for a
timeout, and the API more directly expresses the intent of the test.