Stop holding the InvokerZone on the Invoker
The commented purpose and usage of this field are to avoid problems
where a `heartbeat` call is in a `Zone` which interferes with `Timer`
behavior. `Zone.root` should always have known behavior, or at the least
the `_invokerZone` would have always been a descendent of this root zone
anyway, so it shouldn't introduce any problems that weren't present
before.