Drop two unnecessary getters in Configuration
Using the smallest pattern possible, rather than matching surround
fields, helps readability because the things that behave differently
look different.

Use a public final field over a private final with a public getter for
`coverage`. This is different from other getters in this class in that
it doesn't do null defaulting.

Use a `late final` over `??=` for lazy initialization of the `knownTags`
field. This is different from the `_knownPreset` use case, which
sometimes needs to be set an on an instance directly instead of falling
through to the default initializing expression.