Fix error when meta add unawaited
We have a definition of `unawaited` which is visible from of of the
`test_api/src/` imports to avoid a transitive dep on `pedeantic`. This
causes a static error onces `meta` has a method with the same name.
`pedantic` will be updated to export this method and `test_api` should
do the same until we migrate all usages to the definition from `meta`.
We can't migrate directly to the definition from `meta` since the
version which adds this has a lower SDK bound of `2.9.0` and we aren't
prepared to set our lower bound that high, especially since that SDK is
not yet published. An explicit show on the import of `meta` is backwards
and forward compatible.