Allow wrapping predicates which take dynamic (dart-lang/matcher#150)
The check against `_Predicate<Object>` was intended to be a top type,
but the migration didn't add a `?` so the static type at the call to
`predicate(x)` was `predicate<Object>(x)`. The predicate matcher first
does a type check against it's `T` when matching values, and `null is
Object` returns `false` in an opt-in library. The effect is that a
`Function(dynamic)` gets unnecessarily limited to a `Function(Object)`
and the type check is done artificially by the matcher instead of by the
function itself.

Check against `Object?` instead to make it a null safe top type. Add a
regression test that fails before the fix is applied.