Allow some async work in main with global declarer (#1314)
When a test script is run directly without the test runner creating a
wrapping script there is a global `Declarer` which gets instantiated
during the first interaction with the test framework, such as a call to
`group` or `test`. This would wait a microtask to start running the
declared test cases, but would not work with a `main` that did other
async work before declaring more tests. We can't support arbitrary async
waits because we don't have a consistent signal that the `main` work is
done, so use `pumpEventQueue` as a best effort.