Add Target annotations to annotation classes
This will allow the analyzer to emit diagnostics when the annotations
are used in places where they will not be honored by the test runner.
When used as annotations (as opposed to arguments to `test` or `group`)
the test runner only looks at the library level and all annotations
apply to an entire test suite.