Add a centralized platform import
Towards #962

Make it easier to import the necessary things for writing a custom
platform. This does not expose `executable` so a custom test runner
would need to import at a minimum this and the executable. The symbols
exported here are the ones used by `flutter_tools` in
`flutter_platform.dart` and `test_wrapper.dart`, if other platforms need
other symbols we may need to expand this some.

The import remains a `src` import since these are not yet stable enough
to be fully public. Centralizing and adding `show` clauses will make it
easier to know what is used and for which intent.