Remove dependency on package:http (#1354)
The usage was unnecessary since the library it is used in is not cross
platform. The `package:http` API is a little nicer, add a small extension
to give a similar API here.

Also reuse the HttpClient across both requests rather than create a
separate client for each one which is what the `package:http` top-level
APIs would do.