Make the RemoteListener channel non-nullable
This channel had been `StremChannel<Object?>` primarily because the
`jsonDocument` transformer could emit `null`. Filter out nulls where
they are decoded to keep the rest of the channel non-nullable.

Refactor the `socketChannel` utility method to resolve casting
difficulties with `StreamChannelTransfomer` APIs. It is easier to
separately deal with the `Stream` and `Sink` portions of the channel
because they have separate needs around casting.

Make `RemoteChannel.start` and `serializeSuite` non-nullable channels.