Add types to more parameters (dart-lang/matcher#162)
Find parameters that are implicitly `dynamic` and add a static type of
either `Object?` or in some cases where a non-null value was clearly
required `Object`. There may be missed cases where a non-null value is
required.

Add `as dynamic` casts back in places where we were already relying on
dynamic calls as required to pass static analysis. This has the added
benefit of making those dangerous calls more obvious.