Do not generated stubs for methods from Object.
The Mock class itself overrides hashCode, operator ==, and toString(), and uses some of these methods, so they should not be stubbed on any class which extends Mock. (This is the `if (type.superclass != null)` => `if (!type.superclass.isDartCoreObject)` changes.)

While testing these fixes, I encountered an issue with relative imports from a non-lib/ directory, and fixed that to use `resolver.assetIdForElement(element)`. However, assetIdForElement returns a Future, which doesn't play well deep in all of the Builder code.

So I extracted out the code that resolves asset IDs into a function, _resolveAssetUris, and an Element Visitor, _TypeVisitor. This code gathers all of the types referenced by any public APIs which need to be mocked, and resolves all of the AssetIds, and then passes a mapping to _MockLibraryInfo.

I also added end-to-end tests to test the runtime behavior of generated mocks. I need to add _many_ more tests in this vein.

PiperOrigin-RevId: 349324517