Use collection expansion to add frame info
Expand a map with the frame info into the full message with `...`. This
keeps the structure more clear and makes it obvious that no other side
effects happen.

Refactor the method finding the frame info. Use `firstWhereOrNull` over
a for loop to find the frame matching the suite path. Separate out two
returns to make it clear that there are always `null` values for some
keys in the map, but not others.