Disable stack trace chaining by default (#1459)
Closes #1358

Stack trace chaining has a negative impact on performance, and the VM
has improved async stack trace chaining by default so the value has
dropped significantly on the VM platform. Despite the value on the web,
it will still be better for the majority of users and use cases to run
tests faster by default, and allow for the `--chain-stack-traces` flag
which can be used when debugging a failure is the explicit goal.