Disable stack trace chains by default - for real (#1470)
Changing the default in the arg parser did not change the runner
behavior, and the test that changed was a false positive for the
behavior difference.

Change the default for `chainStackTraces` in the `Metadata` class which
is the real owner of the defaults. Enable stack trace chaining by
default when using `--verbose-trace`, otherwise default to off.

Remove the extra failure output about enabling stack trace chaining for
load "test" failures since they often have no stack trace.

The output is still very verbose - this commit does not reduce the
output to only once at the end of the tests, but that is something we
should consider doing.

Enable stack trace chaining in some tests where the output is checked in
the test.