Drop support for phantomjs entirely (#1472)
Running tests on `phantomjs` has been broken since the full null safety
migration. We have not been running tests for phanomjs on CI. We have
had a warning about the removal of this support for a while, but the
message mentions version 2.0.0 of `test`, which we won't actually wait
for. Since this isn't working and no one has complained about the
warning or the breaking, and since `phantomjs` itself is deprecated and
chrome headless works better, we will drop support entirely.