Make the `message` argument a Map
Closes #1478

The argument to `PlatformPlugin.load` is always a `Map`. The intent is
always to pass it to `deserializeSuite`, where it is unconditionally
cast to `Map`. The loose `Object` argument type does not add any
capabilities.

Changing `PlatformPlugin.load` argument type is not likely to be
breaking since it is not intended for use from outside callers. Changing
the `deserializeSuite` argument is breaking for platform
implementations. The fix is to use the more narrow argument type for
`PlatformPlugin.load`.