Expand doc and variable name for clarity
Capture some of the stuff I figured out about the usage behind these
variables in the code. Add a doc comment explaining the intent. Make the
name very long so that it won't be accidentally used for a second
purpose without us noticing that we need to update the comment too.

Remove the future that never completes part. As far as I can tell, it
isn't an exercised code path today, and if a usage somehow cropped up
it's very unlikely that a non-completing future would be the right
behavior.