Use the root zone to handle expectAsync callbacks
Previously the call to `Invoker.current.removeOutstandingCallbacks`
happened synchronously when the callback was called. After the change to
use `testHandle.mustWaitFor(Future)` the callback would no longer be
called in a `FakeAsync` zone that does not have any further async
progression. Creating the `Completer` in the root zone avoid this issue.