Add test for opted in package with opted out test
There is a bug with the frontend server usage where the first test that
runs determines the soundness for all other tests. If the first test is
opted in to null safety, all tests must be and an opted out test will
fail to load. If the first test is opted out of null safety, all tests
will run in unsound mode.

Add an integration test package which is opted in to null safety at the
package level and has one test opted out and one opted in. Both tests
will only pass if they are run with the expected soundness.