Dart GenerateMocks function parameter fix.
During the validation phase there is a check that rejects any function that has a result type that includes a non-nullable type parameter. This is because the mock has no way to a priori generate a non-null element of type T to return to the caller.

However, this same check is currently applied to the parameters of mockable methods. The same restriction does not need to be imposed on the result type of function-type parameters because the mock method can simply ignore the function parameter. No value needs to exist of the function's return type (the function can return bottom for all we care).

PiperOrigin-RevId: 373416046