Correct docs for the Retry annotation
Fixes #1529

Drop mention of `TestFailure` - there is no distinction between the
types of exceptions which are thrown.

Correct the phrasing to avoid making it seem like the annotation can be
applied to a single test - as an annotation it can only be applied to
the test suite, and for individual groups or tests a named argument is
used instead.