Move testRandomizeOrderingSeed to global level
Resolves an issue discovered in #1539

This configuration was attached to the `SuiteConfig` likely because that
was the easiest way to communicate it through to the `Engine` where
required. This introduced a bug where a more specific `SuiteConfig` for
a tag configured in `dart_test.yaml` would always override the apparent
configuration for those suites with the default, which is unshuffled.
There was no indication when this happeened.

Remove `testRandomizedOrderingSeed` form the `SuiteCConfiguration` class
entirely instead of trying to track down where the inconsistency slips
in.