Wait for isolate exits from VM platform
Internally the test runner is run with
`--mark-main-isolate-as-system-isolate`, so when the main isolate exits
any paused-at-exit non-system isolates are also closed. This makes it
impossible to keep a debugger attached to, for instance, example the
state of the heap following the tests.

Wait for all started isolates to exit when closing the VM platform so
that the outer "system" isolate doesn't forcibly close them before they
are done.