Try logging to stdout
Logging to stderr had seemed to cause problems on windows, but maybe it
just doesn't work on `stderr`. Try and see.

We have a real failure now, but it was on windows so the logging might
not be complete. We do get _some_ output though, which may be coming
from the `--v=1` flag.