add --ignore-timeouts flag (#1635)
Closes https://github.com/dart-lang/test/issues/1634

Adds an `ignoreTimeouts` field on `Suite`, which is read through the invoker to ignore its timeout behavior.

I also changed the `--pause-after-load` behavior slightly so that it implies this new flag instead of `--timeout=none`. This flag more reliably achieves the desired result (not timeouts) as it won't be overridden by local test config.