Stop treating `/packages/` specially
Fixes #1667

The special treatment of `packages` as a directory name was from when
`pub` would create a directory with that name under the package root. I
think the intention was to avoid loading in tests from the packages pub
downloaded.

There are still cases where a directory named `packages` is created and
treated specially by Dart tooling, but those use cases are limited to
the web and should no longer have their `test` directories retained.